algo-viz/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cryptography==42.0.7; python_version >= '3.7'
defusedxml==0.8.0rc2; python_version >= '3.6'
django-cors-headers==4.3.1; python_version >= '3.8'
django-templated-mail==1.1.1
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.2.2; python_version >= '3.8' and python_version < '4.0'
idna==3.7; python_version >= '3.5'
mysqlclient==2.1.0; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python3-openid==3.2.0
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==2.0.0; python_version >= '3.4'
social-auth-app-django==5.4.1; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'

Django~=5.0.6
djangorestframework~=3.15.1